% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_vectors_statistically.R
\name{aitoa.compare.vectors.statistically}
\alias{aitoa.compare.vectors.statistically}
\title{Perform a Statistical Test Comparing Two Vectors}
\usage{
aitoa.compare.vectors.statistically(x, y, alpha)
}
\arguments{
\item{x}{the first vector}

\item{y}{the second vector}

\item{alpha}{the significance threshols}
}
\value{
the result, a list of values\describe{ \item{p}{the p-value computed
  by the test} \item{r}{a character, either\describe{ \item{<}{if \code{x} is
  significantly smaller than \code{y}} \item{>}{if \code{x} is significantly
  larger than \code{y}} \item{?}{if the difference between \code{x} and
  \code{y}, if any, is not significant} }} }
}
\description{
Compare two vectors of samples using a statistical test (the
  two.sided \link[stats]{wilcox.test}, i.e., the Mann-Whitney U test). This
  function suppresses warnings that emerge from ties and repeated values. It
  is not reliable for productive use, but for the context of our book, it
  will suffice for now.
}
