% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_stat_run.R
\name{aitoa.create.stat.run}
\alias{aitoa.create.stat.run}
\title{Generate a Virtual Run based on a Statistic}
\usage{
aitoa.create.stat.run(
  data,
  x.column = c("t", "fes", "f"),
  y.column = c("f", "t", "fes"),
  stat.func = mean,
  make.stairs = TRUE,
  x.min = NA_real_,
  x.max = NA_real_,
  na.replacement = Inf
)
}
\arguments{
\item{data}{the set of runs}

\item{x.column}{the name of the x column}

\item{y.column}{the name of the y column}

\item{stat.func}{the statistic function to be computed}

\item{make.stairs}{should stairs be computed?}

\item{x.min}{an optional minimal x-coordinate, for which a point will be
generated}

\item{x.max}{an optional maximum x corrdinate, for which a point will be
generated}

\item{na.replacement}{a value to be used instead if \code{NA}.}
}
\value{
a matrix whose first column holds the x-coordinates and whose second
  column holds the y-coordinates, i.e., the computed statistics
}
\description{
For a given set of runs, selected x- and y-columns, a new run is
  computed whose x-coordinates are the unique x-coordinates from the original
  run and the corresponding y-coordinates are the specified statistic,
  computed over all the current values of the runs. Useless/duplicated points
  are deleted.
}
