% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distinct_colors.R
\name{aitoa.distinct.colors}
\alias{aitoa.distinct.colors}
\title{Obtain a Set of \code{n} Distinct Colors}
\usage{
aitoa.distinct.colors(n)
}
\arguments{
\item{n}{the number of colors to get}
}
\value{
a character vector of length \code{n} with these colors
}
\description{
Returns a set of \code{n} differently-looking colors.
}
