% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distinct_styles.R
\name{aitoa.distinct.lty}
\alias{aitoa.distinct.lty}
\title{Obtain a Set of Distinct Line Types}
\usage{
aitoa.distinct.lty(n)
}
\arguments{
\item{n}{the number of line types to generate}
}
\value{
a vector of line types
}
\description{
return \code{n} line types
}
