% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distinct_styles.R
\name{aitoa.distinct.pch}
\alias{aitoa.distinct.pch}
\title{Obtain a Set of Distinct Plot Characters}
\usage{
aitoa.distinct.pch(n)
}
\arguments{
\item{n}{the number of plot characters to generate}
}
\value{
a vector of plot characters
}
\description{
return \code{n} plot characters
}
