% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_ellipse.R
\name{aitoa.draw.ellipse}
\alias{aitoa.draw.ellipse}
\title{Draw an Ellipse}
\usage{
aitoa.draw.ellipse(
  x.center = 0,
  y.center = x.center,
  x.radius = 1,
  y.radius = x.radius,
  resolution = 201L,
  ...
)
}
\arguments{
\item{x.center}{the horizontal center}

\item{y.center}{the vertical center}

\item{x.radius}{the horizontal radius}

\item{y.radius}{the vertical radius}

\item{resolution}{the number of points to use to emulate the ellipse}

\item{...}{the parameters to be passed to \link[graphics]{polygon}}
}
\description{
Draw an ellipse around the specified center with the specified radii.
}
