% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_worse_results_from_literature.R
\name{aitoa.find.worse.results.from.literature}
\alias{aitoa.find.worse.results.from.literature}
\title{Find Rows in a Data Set that are Equal or Worse than the Given Data}
\usage{
aitoa.find.worse.results.from.literature(
  data,
  statistics = c("best.f.mean", "best.f.mean"),
  compare.to = .default.load.jssp.results()
)
}
\arguments{
\item{data}{the result data loaded via \link{aitoa.load.end.result.stats}.
This must be for one single algorithm only.}

\item{statistics}{a character vector of statistics to compare}

\item{compare.to}{the results from literature, e.g., obtained via
\code{jsspInstancesAndResults::jssp.results} from our
\href{https://github.com/thomasWeise/jsspInstancesAndResults}{jsspInstancesAndResults}
package}
}
\value{
a list with two elements: \code{data}, the relevant columns from the
  parameter \code{data}, and \code{not.better}, another data frame with the
  literature records that are not better in any statistic on any instances
  than those in \code{data}, or \code{NULL}, if no such record exists
}
\description{
From records with results from literature similar to those used
  in our
  \href{https://github.com/thomasWeise/jsspInstancesAndResults}{jsspInstancesAndResults}
  package, we try to find setups that perform worse than the given statistics
  record.
}
