% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_numbers.R
\name{aitoa.format.counter}
\alias{aitoa.format.counter}
\title{Format an integer counter}
\usage{
aitoa.format.counter(t)
}
\arguments{
\item{t}{the time number}
}
\value{
the formatted string
}
\description{
format an integer counter value
}
