% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_numbers.R
\name{aitoa.format.fes}
\alias{aitoa.format.fes}
\title{Format a Function Evaluation Counter}
\usage{
aitoa.format.fes(t)
}
\arguments{
\item{t}{the FE number}
}
\value{
the formatted FE string
}
\description{
format an FE value
}
