% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_numbers.R
\name{aitoa.format.improvements}
\alias{aitoa.format.improvements}
\title{Format an improvement statistic}
\usage{
aitoa.format.improvements(t)
}
\arguments{
\item{t}{the time number}
}
\value{
the formatted improvement string
}
\description{
format an improvement counter value
}
