% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_numbers.R
\name{aitoa.format.integer.objective.value}
\alias{aitoa.format.integer.objective.value}
\title{Format an Integer-based Objective Value}
\usage{
aitoa.format.integer.objective.value(t)
}
\arguments{
\item{t}{the objective value}
}
\value{
the formatted string
}
\description{
format an integer objective value
}
