% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_numbers.R
\name{aitoa.format.large.number}
\alias{aitoa.format.large.number}
\title{Format a Large Number to a Text}
\usage{
aitoa.format.large.number(v)
}
\arguments{
\item{v}{the number}
}
\value{
the string representation
}
\description{
Transform a large but finite number to a pleasant string.
}
