% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_numbers.R
\name{aitoa.format.number.markdown}
\alias{aitoa.format.number.markdown}
\title{Format a Number for Presentation in Markdown}
\usage{
aitoa.format.number.markdown(
  n,
  finite.preprocessor = identity,
  real.preprocessor = identity,
  integer.preprocessor = real.preprocessor,
  format.large = aitoa.format.large.number,
  format.small = aitoa.format.small.number,
  format.fraction = aitoa.format.fractional.number,
  const.na = "N/A",
  const.p.inf = "$\\\\infty$",
  const.n.inf = "$-\\\\infty$",
  const.0 = "0",
  const.1 = "1",
  const.m1 = "-1"
)
}
\arguments{
\item{n}{the number}

\item{finite.preprocessor}{a function applied to all finite numbers}

\item{real.preprocessor}{a function applied to all real numbers}

\item{integer.preprocessor}{a function applied to all integer numbers}

\item{format.large}{a formatter for large numbers (larger than 10'000'000'000 in absolute value)}

\item{format.small}{a formatter for numbers which are small}

\item{format.fraction}{a formatter for fractional numbers in (-1, 1)}

\item{const.na}{the string constant for NA}

\item{const.p.inf}{the string constant for positive infinity}

\item{const.n.inf}{the string constant for negative infinity}

\item{const.0}{the string constant for 0}

\item{const.1}{the string constant for 1}

\item{const.m1}{the string constant for -1}
}
\value{
a string representation of the number
}
\description{
Format a number such that it can suitably be presented in Markdown text.
}
