% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_numbers.R
\name{aitoa.format.small.integer.objective.value}
\alias{aitoa.format.small.integer.objective.value}
\title{Format a Small Integer-based Objective Value}
\usage{
aitoa.format.small.integer.objective.value(t)
}
\arguments{
\item{f}{the objective value}
}
\value{
the formatted string
}
\description{
format a small integer objective value
}
