% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_numbers.R
\name{aitoa.format.small.number}
\alias{aitoa.format.small.number}
\title{Format a Small Number to a Text}
\usage{
aitoa.format.small.number(v)
}
\arguments{
\item{v}{the number}
}
\value{
the string representation
}
\description{
Transform a small but finite number to a pleasant string.
}
