% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_numbers.R
\name{aitoa.format.time}
\alias{aitoa.format.time}
\title{Format a Time}
\usage{
aitoa.format.time(t)
}
\arguments{
\item{t}{the time number}
}
\value{
the formatted time string
}
\description{
format a time value
}
