% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_to_matrix.R
\name{aitoa.function.to.matrix}
\alias{aitoa.function.to.matrix}
\title{Create the Data Needed for Plotting a Function in 3D}
\usage{
aitoa.function.to.matrix(
  x.min = -5,
  x.max = if (x.min < 0) -x.min else if (x.min == 0) 1 else max((x.min + 1), 1.1 *
    x.min),
  y.min = x.min,
  y.max = x.max,
  n.points.x = 101L,
  n.points.y = n.points.x,
  f = aitoa.rosenbrock
)
}
\arguments{
\item{x.min}{the minimum value of the x-coordinate}

\item{x.max}{the maximum value of the x-corrdinate}

\item{y.min}{the minimum value of the y-coordinate}

\item{y.max}{the maximum value of the y-corrdinate}

\item{n.points.x}{the number of points to generate along the x-axis}

\item{n.points.y}{the number of points to generate along the y-axis}

\item{f}{a function taking a 2-dimensional vector as input and returning the
z-coordinate as output}
}
\value{
a \code{list(x, y, z, grid)}, where \code{x} and \code{y} are
  numerical vectors of \code{x} and \code{y} coordinates and \code{z} is a
  matrix with the \code{z} coordinate for each \code{x-y} pair and
  \code{grid} is a matrix with two columns and one row, one row per
  \code{x-y} pair
}
\description{
Create the \code{x}, \code{y}, and \code{z} data needed to plot
  a function in 3D
}
