% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphic_out.R
\name{aitoa.graphic}
\alias{aitoa.graphic}
\title{Create a Graphic File}
\usage{
aitoa.graphic(
  directory = ".",
  name,
  type = c("svg", "pdf", "eps", "png"),
  width = 7L,
  height = width/.goldenRatio,
  font.size = 10L,
  dpi = 900L,
  skip.if.exists = TRUE,
  bg = "transparent",
  body = { }
)
}
\arguments{
\item{directory}{the directory where the graphic will be stored}

\item{name}{the name chosen for the graphic}

\item{type}{the type of the graphic; supported are \code{svg}, \code{pdf},
\code{eps}, and \code{png}}

\item{width}{the width of the graphic; set to \code{NA} for default (based on
\code{height}, if available)}

\item{height}{the height of the graphic; set to \code{NA} for default (based
on \code{width}, if available)}

\item{font.size}{the default font size}

\item{dpi}{the dots per inch to use if pixel-based rendering is required}

\item{skip.if.exists}{do not create files that already do exist and skip the
diagram creation if so}

\item{bg}{the background color}

\item{body}{the expression to be executed for painting the graphic.}
}
\value{
the path to the newly created graphic
}
\description{
Create a graphic in the given format and write it into the
  given path. First, the proper graphics driver is selected and the output
  file is allocated. Then, the \code{body} expression is executed. This
  expression is expected to draw the graphic's contents. Finally, the
  graphics file is closed and post-processed. The path to the new graphic is
  returned.
}
