% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_jssp_result.R
\name{aitoa.jssp.parse.result}
\alias{aitoa.jssp.parse.result}
\title{Parse the Result of a JSSP Experiment}
\usage{
aitoa.jssp.parse.result(y)
}
\arguments{
\item{y}{the final result of the experiment, as returned by
\code{aitoa.load.result.from.log.file(..)$best.y}}
}
\value{
the list of lists of lists that can be processed by
  \link{aitoa.plot.gantt}.
}
\description{
Parse the final solution of a run in the JSSP experiment into a
  form that can be processed by \link{aitoa.plot.gantt}.
}
\seealso{
\link{aitoa.plot.gantt}, \link{aitoa.load.result.from.log.file}
}
