% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legends.R
\name{aitoa.legend.label}
\alias{aitoa.legend.label}
\title{A Wrapper to Create Axis Labels as Legends}
\usage{
aitoa.legend.label(
  x,
  legend,
  cex = .legend.cex,
  bty = "0",
  bg = .legend.bg,
  box.lwd = 0L,
  seg.len = -0.6,
  y.intersp = 0,
  lwd = 0,
  pch = NA,
  lty = NA,
  pt.lwd = 0,
  pt.cex = 0,
  inset = 0.01,
  ...
)
}
\arguments{
\item{x}{the legend position}

\item{legend}{the legend text}

\item{cex}{the font size factor}

\item{bty}{the box type}

\item{bg}{the background}

\item{box.lwd}{the box line width}

\item{seg.len}{the segment length}

\item{y.intersp}{the y interspacing}

\item{lwd}{the line width}

\item{pch}{the plot character}

\item{lty}{the line type}

\item{pt.lwd}{the point line width}

\item{pt.cex}{the point cex}

\item{inset}{the insets}

\item{...}{other parameters to be passed to \link[graphics]{legend}}
}
\description{
A shorthand with defaults for axis labels drawn as legends.
}
