% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legends.R
\name{aitoa.legend.main}
\alias{aitoa.legend.main}
\title{A Wrapper for the Main Legend}
\usage{
aitoa.legend.main(
  x,
  legend,
  col = "black",
  text.col = col,
  cex = .legend.cex,
  bg = .legend.bg,
  bty = "o",
  box.lwd = 0L,
  inset = 0.005,
  ...
)
}
\arguments{
\item{x}{the legend position}

\item{legend}{the legend text}

\item{col}{the color}

\item{text.col}{the text color}

\item{cex}{the character resize factor}

\item{bg}{the back ground}

\item{bty}{the box type}

\item{box.lwd}{the box line width}

\item{inset}{the insets}
}
\description{
A shorthand with defaults for main legends.
}
