% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_algorithm_dir.R
\name{aitoa.load.algo.dir}
\alias{aitoa.load.algo.dir}
\title{Load a Directory with all Results for one Algorithm}
\usage{
aitoa.load.algo.dir(
  algo.dir,
  keep.columns = c("fes", "t", "f"),
  make.time.unique = FALSE
)
}
\arguments{
\item{algo.dir}{the algorithm directory}

\item{keep.columns}{the columns to keep, any vector containing elements
\code{"t"} (for time), \code{"f"} (for the objective value), and
\code{"fes"} (for the consumed FEs)}

\item{make.time.unique}{should we make the time indices unique (except maybe
for the first and last point)? This makes sense when we want to plot
diagrams over a time axis, as we then have removed redundant points right
away. If \code{make.time.unique==FALSE}, then there may be multiple
improvements at the same time index due to the resolution of the computer
clock (while each improvement will definitely have a unique FE).}
}
\value{
a list of list of data frames, each loaded via
  \link{aitoa.load.inst.dir}, where the names are the instances
}
\description{
Load all the log files in an algorithm directory recursively
}
\seealso{
\link{aitoa.load.inst.dir}
}
