% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_end_result_stats.R
\name{aitoa.load.end.result.stats}
\alias{aitoa.load.end.result.stats}
\title{Load the End Result Statistics from a File}
\usage{
aitoa.load.end.result.stats(file)
}
\arguments{
\item{file}{the file to load}
}
\value{
the data frame with the end results
}
\description{
Load the end result statistics from a file generated by the
  aitoa log processing.
}
