% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_end_results.R
\name{aitoa.load.end.results}
\alias{aitoa.load.end.results}
\title{Load the End Results from a File}
\usage{
aitoa.load.end.results(file)
}
\arguments{
\item{file}{the file to load}
}
\value{
the data frame with the end results
}
\description{
Load the end results from a file generated by the
  aitoa log processing.
}
