% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_instance_dir.R
\name{aitoa.load.inst.dir}
\alias{aitoa.load.inst.dir}
\title{Load an Instance Directory}
\usage{
aitoa.load.inst.dir(
  inst.dir,
  keep.columns = c("fes", "t", "f"),
  make.time.unique = FALSE,
  f.must.be.improving = TRUE,
  max.runs.to.load = NA_integer_
)
}
\arguments{
\item{inst.dir}{the instance directory}

\item{keep.columns}{the columns to keep, any vector containing elements
\code{"t"} (for time), \code{"f"} (for the objective value), and
\code{"fes"} (for the consumed FEs)}

\item{make.time.unique}{should we make the time indices unique (except maybe
for the first and last point)? This makes sense when we want to plot
diagrams over a time axis, as we then have removed redundant points right
away. If \code{make.time.unique==FALSE}, then there may be multiple
improvements at the same time index due to the resolution of the computer
clock (while each improvement will definitely have a unique FE).}

\item{f.must.be.improving}{\code{true} if the logged objective values must be
strictly improving? This is the default way logs are generated by aitoa.
However, you can also create a log where every single sampled solution is
logged, so then you must set \code{f.must.be.improving=FALSE} to load the
data.}

\item{max.runs.to.load}{an optional limit for the maximum number of runs to
load}
}
\value{
a list of data frames, each loaded with \link{aitoa.load.log.file},
where the "names" are the random seeds
}
\description{
Load all the log files in a directory
}
