% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_result_from_log_file.R
\name{aitoa.load.result.from.log.file}
\alias{aitoa.load.result.from.log.file}
\title{Load the End Result from a Log File}
\usage{
aitoa.load.result.from.log.file(file)
}
\arguments{
\item{file}{the file to load}
}
\value{
a list with the following fields:\describe{ \item{file}{the file from
  which the result was loaded} \item{consumed.fes}{the total number of
  consumed function evaluations} \item{last.improvement.fe}{the function
  evaluation where the last improvement took palce} \item{consumed.time}{the
  total consumed time in milliseconds} \item{last.improvement.time}{the time
  when the last improved solution was found} \item{best.f}{the objective
  value of the best solution found} \item{best.y}{the contents of the BEST_Y
  section, i.e., the string representation of the best solution found}
  \item{best.x}{if the BEST_X section is present, then its contents
  (otherwise this key is not present). If this key is present, then it holds
  the string representation of the internal encoding of the best solution
  found}}
}
\description{
Load the end result of a run from a log file.
}
