% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_stat_result.R
\name{aitoa.load.stat.result}
\alias{aitoa.load.stat.result}
\title{Load the End Result of a Run whose Performance Coincidents with a
  given Statistic}
\usage{
aitoa.load.stat.result(
  end.result.stats,
  results.dir,
  algorithm,
  instance,
  statistic = c("best.f.median", "best.f.min", "best.f.q050", "best.f.q159",
    "best.f.q250", "best.f.q750", "best.f.q841", "best.f.q950", "best.f.max",
    "best.f.mean", "success.time.min", "success.time.max", "success.fes.min",
    "success.fes.max")
)
}
\arguments{
\item{end.result.stats}{the data frame with the end result statistics,
obtained via \link{aitoa.load.end.result.stats}}

\item{results.dir}{the directory with the end results}

\item{algorithm}{the algorithm id}

\item{instance}{the instance if}

\item{statistic}{the statistic for which the matching run is sought}
}
\value{
a list with the following fields:\describe{ \item{file}{the file from
  which the result was loaded} \item{consumed.fes}{the total number of
  consumed function evaluations} \item{last.improvement.fe}{the function
  evaluation where the last improvement took palce} \item{consumed.time}{the
  total consumed time in milliseconds} \item{last.improvement.time}{the time
  when the last improved solution was found} \item{best.f}{the objective
  value of the best solution found} \item{best.y}{the contents of the BEST_Y
  section, i.e., the string representation of the best solution found}
  \item{best.x}{if the BEST_X section is present, then its contents
  (otherwise this key is not present). If this key is present, then it holds
  the string representation of the internal encoding of the best solution
  found}\item{setup}{the setup corresponding to the run}\item{statistic}{the
  value of the parameter \code{statistic}}\item{statistic.value}{the value of
  that statistic, which could be different of the performance indicator of
  the run}}
}
\description{
Load the end results of a run whose performance corresponds
  closely to a given statistic. For instance, by default, this function would
  load the end result of the run where the quality of best discovered
  solution is closest to the median best quality over all runs of the given
  algorithm/instance combination. Basically, this method links the setups
  discoverd with \link{aitoa.load.end.result.stats} to the routine
  \code{aitoa.load.result.from.log.file}.
}
