% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_color_transparent.R
\name{aitoa.make.color.transparent}
\alias{aitoa.make.color.transparent}
\title{Make a Color Partially Transparent}
\usage{
aitoa.make.color.transparent(color, transparency = .default.transparency)
}
\arguments{
\item{color}{the color}

\item{transparency}{the degree of transparency: 1 for maximally transparent
(i.e., invisible), 0 for no transparency at all (i.e., the original color)}
}
\description{
Take a color and add some transparency to it.
}
