% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_end_result_test_table.R
\name{aitoa.make.end.result.test.table.md}
\alias{aitoa.make.end.result.test.table.md}
\title{Create a Table Comparing End Results}
\usage{
aitoa.make.end.result.test.table.md(
  end.results,
  algorithms,
  instances,
  alpha = 0.02,
  correction = c("bonferroni", "none"),
  p.value.format = aitoa.format.fractional.number,
  result.char.format = aitoa.format.setup,
  instance.format = aitoa.format.setup,
  algorithm.format = aitoa.format.setup,
  column = c("best.f", "total.time", "total.fes", "last.improvement.time",
    "last.improvement.fes", "n.improvements")
)
}
\arguments{
\item{end.results}{the end result statistics frame}

\item{algorithms}{the list of algorithms, whose \code{names} will be
displayed in the table}

\item{instances}{the list of instances, whose \code{names} will be used}

\item{alpha}{the significance level}

\item{correction}{should we apply a (e.g., bonferroni) correction?}

\item{p.value.format}{the format for printing p-values}

\item{result.char.format}{the format for the result characters}

\item{instance.format}{the format for instance names}

\item{algorithm.format}{the format for algorithm names}

\item{column}{the column to be compared}
}
\description{
Create a table with statistical test results
}
