% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_stat_table.R
\name{aitoa.make.stat.table.md}
\alias{aitoa.make.stat.table.md}
\title{Make a Statistics Table}
\usage{
aitoa.make.stat.table.md(
  end.result.stats,
  algorithms,
  instances,
  stat.cols = list(best = "best.f.min", mean = "best.f.mean", med = "best.f.median", sd
    = "best.f.sd", `med(t)` = "last.improvement.time.median", `med(FEs)` =
    "last.improvement.fe.median"),
  instance.col = if (length(instances) > 1L) "$\\\\inst.id$" else NULL,
  instance.format = aitoa.format.setup,
  instances.limit = NULL,
  instances.limit.col = if (is.null(instances.limit)) NULL else
    "$\\\\lowerBound{\\\\objf}$",
  instances.limit.format = aitoa.format.small.integer.objective.value,
  algorithms.col = if (length(algorithms) > 1L) "setup" else NULL,
  algorithm.format = aitoa.format.setup,
  stat.cols.format = aitoa.statistics.cols.to.formats(stat.cols),
  mark.smallest.stat = rep_len(TRUE, length(stat.cols)),
  smallest.stat.start = "**",
  smallest.stat.end = "**"
)
}
\arguments{
\item{end.result.stats}{the data frame with the statistics}

\item{algorithms}{the list of algorithms, using \code{names} as their names}

\item{instances}{the instances, using \code{names} as their names}

\item{stat.cols}{the statistics columns to plot}

\item{instance.col}{the print name of the instance column}

\item{instance.format}{the format for the instance name}

\item{instances.limit}{an optional vector of instance limits}

\item{instances.limit.col}{an optional title for the instances limit column}

\item{instances.limit.format}{the formatter for the instances limits}

\item{algorithms.col}{the print name of the algorithms column}

\item{algorithm.format}{the algorithm format}

\item{stat.cols.format}{the formats for the statistics columns}

\item{mark.smallest.stat}{should the smallest statistic be marked (per
column)}

\item{smallest.stat.start}{the start string of the smallest stat mark}

\item{smallest.stat.end}{the end string of the smallest stat mark}
}
\value{
a character array with the data in markdown format
}
\description{
Create a table with standard statistics.
}
