% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_file_name.R
\name{aitoa.parse.file.name}
\alias{aitoa.parse.file.name}
\title{Parse a File Name}
\usage{
aitoa.parse.file.name(file)
}
\arguments{
\item{file}{the file name}
}
\value{
a list with the following components: \code{(algorithm, instance,
  seed)}
}
\description{
Accepts the file name of a log file, returns a list with its
  components: \code{(algorithm, instance, seed)}.
}
