% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gantt_for_stat_on_multiple_instances.R
\name{aitoa.plot.gantt.for.stat.on.multiple.instances}
\alias{aitoa.plot.gantt.for.stat.on.multiple.instances}
\title{Plot Multiple Gantt Charts for one Statistic into one Diagram}
\usage{
aitoa.plot.gantt.for.stat.on.multiple.instances(
  end.result.stats,
  results.dir,
  algorithm,
  instances,
  statistic = c("best.f.median", "best.f.min", "best.f.q050", "best.f.q159",
    "best.f.q250", "best.f.q750", "best.f.q841", "best.f.q950", "best.f.max",
    "best.f.mean", "success.time.min", "success.time.max", "success.fes.min",
    "success.fes.max"),
  machine.name.func = as.character,
  job.colors = NA_character_,
  print.job.names = TRUE,
  job.name.func = as.character,
  job.name.cex = .gantt.default.job.name.cex,
  xlab = NA_character_,
  ylab = NA_character_,
  time.max = NA_integer_,
  instances.limit = NA_integer_,
  instances.limit.name = NA_character_,
  instance.limit.cex = .instance.limit.cex,
  instance.limit.color = .instance.limit.color,
  instance.limit.lty = .instance.limit.lty,
  instance.limit.lwd = .instance.limit.lwd,
  instance.limit.adj = .gantt.default.instance.limit.adj,
  center.label.cex = .gantt.label.cex,
  center.label.bg = .gantt.label.bg,
  las = 1L,
  xaxs = "i",
  yaxs = "i",
  mgp = .default.mgp,
  tck = .default.tck,
  cex = .default.cex,
  mar = if ((is.null(xlab) || is.na(xlab)) && (is.null(ylab) || is.na(ylab)))
    .default.mar.without.labels else NULL,
  mar.single = .default.mar.without.labels,
  ...
)
}
\arguments{
\item{end.result.stats}{the end result statistics frame, as obtained with
\link{aitoa.load.end.result.stats}}

\item{results.dir}{the directory where to search the run log files}

\item{algorithm}{the algorithm id}

\item{instances}{the instance ids, the names of the list are used as instance
names}

\item{statistic}{the statistics for which a matching run should be sought}

\item{machine.name.func}{a function converting a machine index into a
character string with the machine name. The first machine index is
\code{0L}, the last one \code{length(x) - 1L}.}

\item{job.colors}{an optional vector of job colors}

\item{print.job.names}{should the job names be printed into the job
rectangles?}

\item{job.name.func}{a function converting a job index into a character
string, only used if \code{isTRUE(print.job.names)}, the job indices passed
on depend on the indices present in \code{x}.}

\item{job.name.cex}{the scaling for job names, only considered if
\code{isTRUE(print.job.names)}}

\item{xlab}{the label for the x-axis}

\item{ylab}{the label for the y-axis}

\item{time.max}{an optional maximal time value}

\item{instances.limit}{an opional quality limit to be plotted as horizontal
line}

\item{instances.limit.name}{the optional name of the quality limit, \code{NA}
for omit}

\item{instance.limit.cex}{the font scaling for the instance limit annotation
(only if \code{!is.na(instance.limit.name)})}

\item{instance.limit.color}{the color for the instance limit line}

\item{instance.limit.lty}{the line type for the instance limit line}

\item{instance.limit.lwd}{the line width for the instance limit line}

\item{instance.limit.adj}{the adjustment for the instance limit  annotation
(only if \code{!is.na(instance.limit.name)})}

\item{center.label.cex}{the font size multiplier for the label}

\item{center.label.bg}{the background for the label}

\item{las}{the axis label oriantation (this is the default to be passed to
\link[graphics]{plot}, better don't change)}

\item{xaxs}{the x-axis interval type (this is the default to be passed to
\link[graphics]{plot}, better don't change)}

\item{yaxs}{the y-axis interval type (this is the default to be passed to
\link[graphics]{plot}, better don't change)}

\item{mgp}{the mgp parameter to be passed to \link[graphics]{plot}}

\item{tck}{the tck parameter to be passed to \link[graphics]{plot}}

\item{cex}{the default character scaling}

\item{mar}{the default margins}

\item{mar.single}{potential individual margins to be passed to the single
diagrams}

\item{...}{parameters to be passed to \link[graphics]{plot}}
}
\description{
Multiple Gantt charts fitting to one statistic are plotted into
  a single diagram in a column. #' @title Plot a Gantt Chart based on End
  Result Statistics

Plot a Gantt chart extract from a log file matching a given end
  result statistic.
}
