% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_progress_stat_on_multiple_instances.R
\name{aitoa.plot.progress.stat.on.multiple.instances}
\alias{aitoa.plot.progress.stat.on.multiple.instances}
\title{Plot the Progress of a Statistic for a Set of Algorithms on a Set of Problem
  Instances}
\usage{
aitoa.plot.progress.stat.on.multiple.instances(
  results.dir,
  algorithms,
  instances,
  time.column = c("t", "fes"),
  max.time = NA_integer_,
  algorithm.colors = aitoa.distinct.colors(length(algorithms)),
  quantiles = .stat.plot.default.quantiles,
  center.stat = median,
  center.lty = .default.lty,
  center.lwd = .thick.lwd,
  quantile.transparency = 0.8,
  instances.limit = NA_integer_,
  instances.limit.name = NULL,
  instance.limit.color = .instance.limit.color,
  instance.limit.lty = .instance.limit.lty,
  instance.limit.lwd = .instance.limit.lwd,
  legend.cex = .legend.cex,
  legend.bg = .legend.bg,
  time.axis.text = if (time.column[[1L]] == "t") .time.ms.text else .time.fes.text,
  quality.axis.text = .quality.text,
  make.time.unique = (time.column[[1L]] == "t"),
  use.f.range.from.raw.data = FALSE,
  mgp = .default.mgp,
  tck = .default.tck,
  cex = .default.cex,
  mar = .default.mar.without.labels,
  mar.single = .default.mar.without.labels,
  ...
)
}
\arguments{
\item{results.dir}{the directory where the results can be loaded from}

\item{algorithms}{the list of algorithm IDs. The \code{names} of this list,
if set, will be used in the legends.}

\item{instances}{the list of instance IDs. The \code{names} of this list, if
set, will be used in the legend}

\item{time.column}{the time dimension, either \code{t} or \code{fes}}

\item{max.time}{an optional limit for the time, can be a vector}

\item{algorithm.colors}{a character vector of the same length as
\code{algorithms} providing the colors to be used for the algorithms}

\item{quantiles}{the quantiles to be plotted in a shaded fashion: all values
must be >= 0 and < 0.5 and are mirrored. I.e., if you specify 0.2, then the
region between the 0.2 and 0.8 quantile will be plotted semi-transparent.}

\item{center.stat}{the central statistic to plot, usually the \link{median}
or \link{mean}}

\item{center.lty}{the line type to be used for the median lines}

\item{center.lwd}{the line width to be used for median lines}

\item{quantile.transparency}{the transparency to be applied to each quantile
polygon}

\item{instances.limit}{an optional vector of lower bounds or best-known
solutions for the instances}

\item{instances.limit.name}{an optional name or name vector for the instances
limits}

\item{instance.limit.color}{the color for the instance limit line}

\item{instance.limit.lty}{the line type for the instance limit line}

\item{instance.limit.lwd}{the line width for the instance limit line}

\item{legend.cex}{the character scaling for the legend}

\item{legend.bg}{the background color for the legend}

\item{time.axis.text}{the text to be used for labeling the time axis,
\code{NA} for omit label}

\item{quality.axis.text}{the text to be used for labeling the quality axis,
\code{NA} for omit label}

\item{make.time.unique}{should we make the time indices unique (except maybe
for the first and last point)? This makes sense when we want to plot
diagrams over a time axis, as we then have removed redundant points right
away. If \code{make.time.unique==FALSE}, then there may be multiple
improvements at the same time index due to the resolution of the computer
clock (while each improvement will definitely have a unique FE).}

\item{use.f.range.from.raw.data}{should we use the real data to compute the range of the y-axis, or should the range depend on the computed statistics only (default)?}

\item{mgp}{the mgp parameter to be passed to \link[graphics]{plot}}

\item{tck}{the tck parameter to be passed to \link[graphics]{plot}}

\item{cex}{the default character scaling}

\item{mar}{the default margins, can be a list of vectors, one for each plot}

\item{mar.single}{the margin parameter(s) to be passed to the single diagrams}

\item{...}{parameters to be passed to \link[graphics]{par}}
}
\description{
Plot how a statistic for each element of set of algorithms progress over a set of problem
  instances. For each instance, one diagram is plotted. The diagrams are
  arranged one by one from in a vertical row.
}
