% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_runs_on_instance.R
\name{aitoa.plot.runs.on.instance}
\alias{aitoa.plot.runs.on.instance}
\title{Plot the Progress of Several Runs one Algorithm on One Instance}
\usage{
aitoa.plot.runs.on.instance(
  results.dir = ".",
  algorithm,
  instance,
  time.column = c("t", "fes"),
  max.time = NA_integer_,
  run.colors = aitoa.distinct.colors,
  run.lty = .default.lty,
  run.lwd = .default.lwd,
  make.time.unique = FALSE,
  f.must.be.improving = TRUE,
  max.runs.to.plot = NA_integer_,
  legend.cex = .legend.cex,
  legend.bg = .legend.bg,
  time.axis.text = if (time.column == "t") .time.ms.text else .time.fes.text,
  quality.axis.text = .quality.text,
  mgp = .default.mgp,
  tck = .default.tck,
  cex = .default.cex,
  mar = .default.mar.without.labels,
  ...
)
}
\arguments{
\item{results.dir}{the directory where the results can be loaded from}

\item{algorithm}{the algorithm id}

\item{instance}{the instance id to be plotted}

\item{time.column}{the time column}

\item{max.time}{an optional maximal time limit}

\item{run.colors}{a function to generate the colors for the runs}

\item{run.lty}{the line type to be used for the runs, can be
vector}

\item{run.lwd}{the line width to be used for the runs, can be
vector}

\item{make.time.unique}{should we make the time indices unique (except maybe
for the first and last point)? This makes sense when we want to plot
diagrams over a time axis, as we then have removed redundant points right
away. If \code{make.time.unique==FALSE}, then there may be multiple
improvements at the same time index due to the resolution of the computer
clock (while each improvement will definitely have a unique FE).}

\item{f.must.be.improving}{\code{true} if the logged objective values must be
strictly improving? This is the default way logs are generated by aitoa.
However, you can also create a log where every single sampled solution is
logged, so then you must set \code{f.must.be.improving=FALSE} to load the
data.}

\item{max.runs.to.plot}{an optional number giving the maximum number of runs
to be plotted per algorithm}

\item{legend.cex}{the character scaling for the legend}

\item{legend.bg}{the background color for the legend}

\item{time.axis.text}{the text to be used for labeling the time axis,
\code{NA} for omit label}

\item{quality.axis.text}{the text to be used for labeling the quality axis,
\code{NA} for omit label}

\item{mgp}{the mgp parameter to be passed to \link[graphics]{plot}}

\item{tck}{the tck parameter to be passed to \link[graphics]{plot}}

\item{cex}{the default character scaling}

\item{mar}{the default margins}

\item{...}{parameters to be passed to \link[graphics]{par}}
}
\description{
Plot the progress of a set of runs of one algorithm on one instance.
}
