% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stat_over_param.R
\name{aitoa.plot.stat.over.param}
\alias{aitoa.plot.stat.over.param}
\title{Plot a Statistic over the Values of a Given Parameter}
\usage{
aitoa.plot.stat.over.param(
  end.result.stats,
  algorithm.template,
  algorithm.primary.args,
  algorithm.secondary.args = NULL,
  algorithm.secondary.args.name = NULL,
  instances,
  statistic,
  instance.colors = aitoa.distinct.colors(length(instances)),
  instance.lwd = .default.lwd,
  instance.pch = NULL,
  secondary.arg.lty = if (is.null(algorithm.secondary.args)) .default.lty else c(1L,
    aitoa.distinct.lty(length(algorithm.secondary.args) - 1L)),
  legend.pos = "topright",
  legend.cex = .legend.cex,
  legend.bg = .legend.bg,
  mgp = .default.mgp,
  tck = .default.tck,
  cex = .default.cex,
  mar = .default.mar.without.labels,
  divide.by = NULL,
  mark.min.fun = .default.mark.min,
  x.axis.at = NULL,
  ...
)
}
\arguments{
\item{end.result.stats}{the end result statistics frame}

\item{algorithm.template}{the algorithm template: a string where \code{$arg}
will be replaced with the parameter value and which will then be matched
with the \code{algorithm} column in the data frame}

\item{algorithm.primary.args}{the values of the primary argument}

\item{algorithm.secondary.args}{values of an optional secondary argument}

\item{algorithm.secondary.args.name}{the optional name of the optional secondary argument}

\item{instances}{a potentially named list of instances; if names are given,
the names are diplayed in the legend, otherwise the instance strings}

\item{statistic}{the statistic to plot}

\item{instance.colors}{the colors to use for the instances}

\item{instance.lwd}{the line widths to be used for the instances}

\item{instance.pch}{the symbol to be used for the instances}

\item{secondary.arg.lty}{the line types to be used for the secondary argument}

\item{legend.pos}{the legend position}

\item{legend.cex}{the legend character size}

\item{legend.bg}{the legend background}

\item{mgp}{the mgp value}

\item{tck}{the tck value}

\item{cex}{the base character scaling}

\item{mar}{the margin value}

\item{divide.by}{optionally, a per-instance value to divide the statistic by}

\item{mark.min.fun}{the function for marking the minima. If not \code{NULL}, this must be a function with six arguments, namely\describe{
\item{min.x}{the vector with the x-coordinates minal values}
\item{min.y}{the vector with the y-coordinates minal values}
\item{color}{the color for these points}
\item{pch}{the pch value for these points, or \code{NULL} if none}
\item{lty}{the line type for these points}
\item{lwd}{the line width for these points}}
This function, if not \code{NULL}, is called once for each instance}

\item{x.axis.at}{the location of the x-axis ticks: optional}

\item{...}{parameters to be passed to \link[graphics]{plot}}
}
\description{
Plot an end statistic over the values of a given parameter.
}
