% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stat_over_instance.R
\name{aitoa.plot.stats.over.instance}
\alias{aitoa.plot.stats.over.instance}
\title{Plot a Statistic for Different Algorthms over Instances}
\usage{
aitoa.plot.stats.over.instance(
  end.result.stats,
  algorithms,
  instance.selector = function(n) TRUE,
  instance.scaler = as.integer,
  statistic = "last.improvement.fes.mean",
  statistic.top = "last.improvement.fes.q841",
  statistic.bottom = "last.improvement.fes.q159",
  algorithms.color = aitoa.distinct.colors(length(algorithms)),
  algorithms.lty = aitoa.distinct.lty(length(algorithms)),
  algorithms.pch = aitoa.distinct.pch(length(algorithms)),
  algorithms.lwd = .default.lwd,
  statistics.bg.transparency = .default.transparency,
  legend.pos = "topleft",
  legend.cex = .legend.cex,
  legend.bg = .legend.bg,
  include.bg.stats.in.legend = TRUE,
  mgp = .default.mgp,
  tck = .default.tck,
  cex = .default.cex,
  mar = .default.mar.without.labels,
  ...
)
}
\arguments{
\item{end.result.stats}{the end results statistics record}

\item{algorithms}{the algorithms to plot, where \code{names(algorithms)} are
the names, if provided}

\item{instance.selector}{a function which returns \code{TRUE} of each
instance to be included in the plot}

\item{instance.scaler}{a function that returns the x-coordinate for a
selected instance}

\item{statistic}{the statistic to plot}

\item{statistic.top}{a second statistic which should be bigger than
\code{statistic} and together with \code{statistic.bottom} is used to shade
a region behind the \code{statistic}}

\item{statistic.bottom}{a second statistic which should be smaller than
\code{statistic} and together with \code{statistic.top} is used to shade a
region behind the \code{statistic}}

\item{algorithms.color}{the colors to be used for the algorithms}

\item{algorithms.lty}{the line types to be used for the algorithms}

\item{algorithms.pch}{the plot characters to be used for the algorithms}

\item{algorithms.lwd}{the line-widths to be used for the algorithms}

\item{statistics.bg.transparency}{the transparency value (0 for transparent,
1 for normal color) used for making the region behind the statistics
transparent}

\item{legend.pos}{the position for the legend}

\item{legend.cex}{the font size for the legend}

\item{legend.bg}{the background for the legend}

\item{include.bg.stats.in.legend}{should be include the background statistics
into the legend?}

\item{mgp}{the mgp parameter to be passed to \link[graphics]{plot}}

\item{tck}{the tck parameter to be passed to \link[graphics]{plot}}

\item{cex}{the default character scaling}

\item{mar}{the default margins}

\item{...}{parameters to be passed to \link[graphics]{par}}
}
\description{
For each instance, we can compute one x-coordinate. We than plot
  the given statistics of the specified algorithms over this coordinate.
}
