% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_by_stats.R
\name{aitoa.rank.by.stats}
\alias{aitoa.rank.by.stats}
\title{Rank Algorithms by Statistics over Instances}
\usage{
aitoa.rank.by.stats(
  end.result.stats,
  algorithms,
  instances,
  stat.cols = c("best.f.min", "best.f.mean", "best.f.median")
)
}
\arguments{
\item{end.result.stats}{the data frame with the statistics}

\item{algorithms}{the list of algorithms}

\item{instances}{the instances}

\item{stat.cols}{the statistics columns to use for ranking}
}
\value{
a data frame with the ranks
}
\description{
Rank a set of algorithms over a set of instances by their statstics and return a data frame.
}
