% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_to_matrix.R
\name{aitoa.rastrigin}
\alias{aitoa.rastrigin}
\title{The 2-dimensional Rastrigin Function}
\usage{
aitoa.rastrigin(x)
}
\arguments{
\item{x}{the 2-dimensional input vector}
}
\value{
the value of the Rastrigin function
}
\description{
The 2-dimensional Rastrigin Function
}
