% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgb_mix.R
\name{aitoa.rgb.mix}
\alias{aitoa.rgb.mix}
\title{Mix Two RGB Colors}
\usage{
aitoa.rgb.mix(
  color.1,
  color.2,
  weight = 0.5,
  limit = 255L,
  make.int = (limit > 1L)
)
}
\arguments{
\item{color.1}{the first r-g-b vector}

\item{color.2}{the second r-g-b vector}

\item{weight}{the weight of the first color vector}

\item{limit}{the color limit}

\item{make.int}{should we convert the result to an integer?}
}
\value{
the mixed color
}
\description{
Mix two RGB colors
}
