% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_to_matrix.R
\name{aitoa.rosenbrock}
\alias{aitoa.rosenbrock}
\title{The 2-dimensional Rosenbrock Function}
\usage{
aitoa.rosenbrock(x)
}
\arguments{
\item{x}{the 2-dimensional input vector}
}
\value{
the value of the Rosenbrock function
}
\description{
The 2-dimensional Rosenbrock Function
}
