% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated_annealing_temperature.R
\name{aitoa.sa.generate.tau}
\alias{aitoa.sa.generate.tau}
\title{Generate a Sequence of Tau values}
\usage{
aitoa.sa.generate.tau(max.tau = 30000000L)
}
\arguments{
\item{max.tau}{the maximum tau}
}
\value{
the sequence
}
\description{
Generate a sequence of time steps
}
