% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated_annealing_temperature.R
\name{aitoa.sa.P}
\alias{aitoa.sa.P}
\title{Compute the Simulated Annealing Acceptance Probability}
\usage{
aitoa.sa.P(deltaE, T)
}
\arguments{
\item{deltaE}{the single difference in objective value}

\item{T}{the vector of temperatures}
}
\value{
the acceptance probabilities
}
\description{
Compute the Acceptance Probability for a given deltaE and
  temperature T.
}
