% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated_annealing_temperature.R
\name{aitoa.sa.T.exp}
\alias{aitoa.sa.T.exp}
\title{The Simulated Annealing Exponential Temperature Schedule}
\usage{
aitoa.sa.T.exp(Ts, epsilon, tau)
}
\arguments{
\item{Ts}{the start temperature}

\item{epsilon}{the epsilon}

\item{tau}{the integer vector of FE indices}
}
\value{
the temperatures
}
\description{
Compute the temperature at FE index tau
}
