% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated_annealing_temperature.R
\name{aitoa.sa.T.log}
\alias{aitoa.sa.T.log}
\title{The Simulated Annealing Logarithmic Temperature Schedule}
\usage{
aitoa.sa.T.log(Ts, epsilon = 1, tau)
}
\arguments{
\item{Ts}{the start temperature}

\item{epsilon}{the epsilon}

\item{tau}{the integer vector of FE indices}
}
\value{
the temperatures
}
\description{
Compute the temperature at FE index tau
}
