% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated_annealing_temperature.R
\name{aitoa.sa.temperature.plot}
\alias{aitoa.sa.temperature.plot}
\title{Create the Temperature/Probability Plot for Simulated Annealing}
\usage{
aitoa.sa.temperature.plot()
}
\description{
This is not really a function for evaluating stuff, but it
  creates a nice plot that I use in the book, so I want to keep it.
}
