% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_function.R
\name{aitoa.sample.function}
\alias{aitoa.sample.function}
\title{Sample a Function at Some Points in the Given Interval}
\usage{
aitoa.sample.function(f, x.min, x.max, keep.only = is.finite)
}
\arguments{
\item{f}{the function to sample}

\item{x.min}{the minimum (inclusive) x value}

\item{x.max}{the maximum (inclusive) x value}

\item{keep.only}{a selector: only points meeting this specification will be
returned}
}
\value{
a list with the x and y coordinates of the samples
}
\description{
Sample a function so that it can be plotted smoothly.
}
