% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_numbers.R
\name{aitoa.statistics.cols.to.formats}
\alias{aitoa.statistics.cols.to.formats}
\title{Convert a Column Type to a Formatter Function}
\usage{
aitoa.statistics.cols.to.formats(
  cols,
  format.time = aitoa.format.time,
  format.fes = aitoa.format.fes,
  format.improvements = aitoa.format.improvements,
  format.objective = aitoa.format.small.integer.objective.value,
  format.counter = aitoa.format.counter,
  format.setup = aitoa.format.setup
)
}
\arguments{
\item{cols}{the column types}

\item{format.time}{the time formatting function}

\item{format.fes}{the FEs formatting function}

\item{format.improvements}{the improvements formatting function}

\item{format.objective}{the objective formatting function}

\item{format.counter}{the counter formatting function}

\item{format.setup}{the setup formatting function}
}
\value{
the formatter function
}
\description{
Convert a Column Type to a Formatter Function
}
