% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_out.R
\name{aitoa.text}
\alias{aitoa.text}
\title{Create a Text File}
\usage{
aitoa.text(
  directory = ".",
  name,
  type = c("md", "txt"),
  trim.ws = TRUE,
  skip.if.exists = TRUE,
  body = { }
)
}
\arguments{
\item{directory}{the directory where the text will be stored}

\item{name}{the name chosen for the text}

\item{type}{the type of the graphic; supported are \code{txt} and \code{md}}

\item{trim.ws}{should whitespace be trimmed?}

\item{skip.if.exists}{should existing files be skipped?}

\item{body}{the expression to be executed for getting the text.}
}
\value{
the path to the newly created text file
}
\description{
Create a text file.
}
