% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_algorithm_dir.R
\name{aitoa.load.algo.dir}
\alias{aitoa.load.algo.dir}
\title{Load a Directory with all Results for one Algorithm}
\usage{
aitoa.load.algo.dir(algoDir, keepColumns = .default.colums,
  makeTimeUnique = FALSE)
}
\arguments{
\item{algoDir}{the algorithm directory}

\item{keepColumns}{the columns to keep, any vector containing elements
\code{"t"} (for time), \code{"f"} (for the objective value), and
\code{"fes"} (for the consumed FEs)}

\item{makeTimeUnique}{should we make the time indices unique (except maybe
for the first and last point)? This makes sense when we want to plot
diagrams over a time axis, as we then have removed redundant points right
away. If \code{makeTimeUnique==FALSE}, then there may be multiple
improvements at the same time index due to the resolution of the computer
clock (while each improvement will definitely have a unique FE).}
}
\value{
a list of list of data frames, where the names are the instance IDs
}
\description{
Load all the log files in an algorithm directory recursively
}
