% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_instance_dir.R
\name{aitoa.load.inst.dir}
\alias{aitoa.load.inst.dir}
\title{Load an Instance Directory}
\usage{
aitoa.load.inst.dir(
  inst.dir,
  keep.columns = c("fes", "t", "f"),
  make.time.unique = FALSE
)
}
\arguments{
\item{inst.dir}{the instance directory}

\item{keep.columns}{the columns to keep, any vector containing elements
\code{"t"} (for time), \code{"f"} (for the objective value), and
\code{"fes"} (for the consumed FEs)}

\item{make.time.unique}{should we make the time indices unique (except maybe
for the first and last point)? This makes sense when we want to plot
diagrams over a time axis, as we then have removed redundant points right
away. If \code{make.time.unique==FALSE}, then there may be multiple
improvements at the same time index due to the resolution of the computer
clock (while each improvement will definitely have a unique FE).}
}
\value{
a list of data frames, each loaded with \link{aitoa.load.log.file},
where the "names" are the random seeds
}
\description{
Load all the log files in a directory
}
