% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gantt.R
\name{aitoa.plot.gantt}
\alias{aitoa.plot.gantt}
\title{Plot a Gantt Chart}
\usage{
aitoa.plot.gantt(
  x,
  machine.name.prefix = NA_character_,
  machine.name.start = 0L,
  job.colors = NA_character_,
  print.job.names = TRUE,
  job.name.prefix = NA_character_,
  job.name.start = NA_integer_,
  job.name.cex = 0.9,
  xlab = "Time",
  ylab = "Machine",
  las = 1L,
  xaxs = "i",
  yaxs = "i",
  ...
)
}
\arguments{
\item{x}{the data to be plotted}

\item{machine.name.prefix}{an optional prefix for machine names}

\item{machine.name.start}{the starting index for machine names}

\item{job.colors}{an optional vector of job colors}

\item{print.job.names}{should the job names be printed into the job
rectangles?}

\item{job.name.prefix}{an optional prefix for job names, only considered if
\code{isTRUE(print.job.names)}}

\item{job.name.start}{the index assigned to the lowest job ID, only
considered if \code{isTRUE(print.job.names)}}

\item{job.name.cex}{the scaling for job names, only considered if
\code{isTRUE(print.job.names)}}

\item{xlab}{the label for the x-axis}

\item{ylab}{the label for the y-axis}

\item{las}{the axis label oriantation (this is the default to be passed to
\link[graphics]{plot}, better don't change)}

\item{xaxs}{the x-axis interval type (this is the default to be passed to
\link[graphics]{plot}, better don't change)}

\item{yaxs}{the y-axis interval type (this is the default to be passed to
\link[graphics]{plot}, better don't change)}

\item{...}{parameters to be passed to \link[graphics]{plot}}
}
\description{
Plot a Gantt chart based on a list \code{x} of lists of data.
  The list contains one list for each machine. Each machine list, in turn, is
  a list of lists, too. Each of their elements has the form \code{list(job=,
  start=, end=)}, with elements denoting the job ID, start, and end time,
  respectively.
}
