% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gantt.R
\name{aitoa.plot.gantt}
\alias{aitoa.plot.gantt}
\title{Plot a Gantt Chart}
\usage{
aitoa.plot.gantt(
  x,
  machine.name.func = as.character,
  job.colors = NA_character_,
  print.job.names = TRUE,
  job.name.func = as.character,
  job.name.cex = .gantt.default.job.name.cex,
  xlab = NA_character_,
  ylab = NA_character_,
  time.max = NA_integer_,
  instance.limit = NA_integer_,
  instance.limit.name = NA_character_,
  instance.limit.cex = .instance.limit.cex,
  instance.limit.color = .instance.limit.color,
  instance.limit.lty = .instance.limit.lty,
  instance.limit.lwd = .instance.limit.lwd,
  instance.limit.adj = .gantt.default.instance.limit.adj,
  center.label = NA_character_,
  center.label.cex = .gantt.label.cex,
  center.label.bg = .gantt.label.bg,
  las = 1L,
  xaxs = "i",
  yaxs = "i",
  mgp = .default.mgp,
  tck = .default.tck,
  cex = .default.cex,
  mar = if ((is.null(xlab) || is.na(xlab)) && (is.null(ylab) || is.na(ylab)))
    .default.mar.without.labels else NULL,
  ...
)
}
\arguments{
\item{x}{the data to be plotted}

\item{machine.name.func}{a function converting a machine index into a
character string with the machine name. The first machine index is
\code{0L}, the last one \code{length(x) - 1L}.}

\item{job.colors}{an optional vector of job colors}

\item{print.job.names}{should the job names be printed into the job
rectangles?}

\item{job.name.func}{a function converting a job index into a character
string, only used if \code{isTRUE(print.job.names)}, the job indices passed
on depend on the indices present in \code{x}.}

\item{job.name.cex}{the scaling for job names, only considered if
\code{isTRUE(print.job.names)}}

\item{xlab}{the label for the x-axis}

\item{ylab}{the label for the y-axis}

\item{time.max}{an optional maximal time value}

\item{instance.limit}{an opional quality limit to be plotted as horizontal
line}

\item{instance.limit.name}{the optional name of the quality limit, \code{NA}
for omit}

\item{instance.limit.cex}{the font scaling for the instance limit annotation
(only if \code{!is.na(instance.limit.name)})}

\item{instance.limit.color}{the color for the instance limit line}

\item{instance.limit.lty}{the line type for the instance limit line}

\item{instance.limit.lwd}{the line width for the instance limit line}

\item{instance.limit.adj}{the adjustment for the instance limit  annotation
(only if \code{!is.na(instance.limit.name)})}

\item{center.label}{an optional label to be plotted in the lower center}

\item{center.label.cex}{the font size multiplier for the label}

\item{center.label.bg}{the background for the label}

\item{las}{the axis label oriantation (this is the default to be passed to
\link[graphics]{plot}, better don't change)}

\item{xaxs}{the x-axis interval type (this is the default to be passed to
\link[graphics]{plot}, better don't change)}

\item{yaxs}{the y-axis interval type (this is the default to be passed to
\link[graphics]{plot}, better don't change)}

\item{mgp}{the mgp parameter to be passed to \link[graphics]{plot}}

\item{tck}{the tck parameter to be passed to \link[graphics]{plot}}

\item{cex}{the default character scaling}

\item{mar}{the default margins}

\item{...}{parameters to be passed to \link[graphics]{plot}}
}
\description{
Plot a Gantt chart based on a list \code{x} of lists of data.
  The list contains one list for each machine. Each machine list, in turn, is
  a list of lists, too. Each of their elements has the form \code{list(job=,
  start=, end=)}, with elements denoting the job ID, start, and end time,
  respectively.
}
