% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_progress_on_instance.R
\name{aitoa.plot.progress.on.instance}
\alias{aitoa.plot.progress.on.instance}
\title{Plot the Progress of a Set of Algorithms on One Instance}
\usage{
aitoa.plot.progress.on.instance(
  results.dir = ".",
  algorithms,
  instance,
  instance.name = instance,
  time.column = c("t", "fes"),
  max.time = NA_integer_,
  algorithm.colors = aitoa.distinct.colors(length(algorithms)),
  algorithm.lty = .default.lty,
  algorithm.lwd = .default.lwd,
  instance.limit = NA_integer_,
  instance.limit.name = NA_character_,
  instance.limit.color = .instance.limit.color,
  instance.limit.lty = .instance.limit.lty,
  instance.limit.lwd = .instance.limit.lwd,
  legend.cex = .legend.cex,
  legend.bg = .legend.bg,
  time.axis.text = if (time.column[[1L]] == "t") .time.ms.text else .time.fes.text,
  quality.axis.text = .quality.text,
  make.time.unique = FALSE,
  f.must.be.improving = TRUE,
  max.runs.to.plot = NA_integer_,
  mgp = .default.mgp,
  tck = .default.tck,
  cex = .default.cex,
  mar = .default.mar.without.labels,
  execute = { },
  ...
)
}
\arguments{
\item{results.dir}{the directory where the results can be loaded from}

\item{algorithms}{a list of algorithsm, the \code{names} of which (if
provided) are used for the legend}

\item{instance}{the instance id to be plotted}

\item{instance.name}{the name of the instance to show in the legend}

\item{time.column}{the time column}

\item{max.time}{an optional maximal time limit}

\item{algorithm.colors}{a character vector of the same length as
\code{algorithms} providing the colors to be used for the algorithms}

\item{algorithm.lty}{the line type to be used for the algorithms, can be
vector}

\item{algorithm.lwd}{the line width to be used for the algorithms, can be
vector}

\item{instance.limit}{an opional quality limit to be plotted as horizontal
line}

\item{instance.limit.name}{the optional name of the quality limit}

\item{instance.limit.color}{the color for the instance limit line}

\item{instance.limit.lty}{the line type for the instance limit line}

\item{instance.limit.lwd}{the line width for the instance limit line}

\item{legend.cex}{the character scaling for the legend}

\item{legend.bg}{the background color for the legend}

\item{time.axis.text}{the text to be used for labeling the time axis,
\code{NA} for omit label}

\item{quality.axis.text}{the text to be used for labeling the quality axis,
\code{NA} for omit label}

\item{make.time.unique}{should we make the time indices unique (except maybe
for the first and last point)? This makes sense when we want to plot
diagrams over a time axis, as we then have removed redundant points right
away. If \code{make.time.unique==FALSE}, then there may be multiple
improvements at the same time index due to the resolution of the computer
clock (while each improvement will definitely have a unique FE).}

\item{f.must.be.improving}{\code{true} if the logged objective values must be
strictly improving? This is the default way logs are generated by aitoa.
However, you can also create a log where every single sampled solution is
logged, so then you must set \code{f.must.be.improving=FALSE} to load the
data.}

\item{max.runs.to.plot}{an optional number giving the maximum number of runs
to be plotted per algorithm}

\item{mgp}{the mgp parameter to be passed to \link[graphics]{plot}}

\item{tck}{the tck parameter to be passed to \link[graphics]{plot}}

\item{cex}{the default character scaling}

\item{mar}{the default margins}

\item{execute}{an arbitrary block of code to be executed within the plotting process}

\item{...}{parameters to be passed to \link[graphics]{par}}
}
\description{
Plot the progress of a set of algorithms on one instance.
}
