% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_progress_on_instance.R
\name{aitoa.plot.progress.on.instance}
\alias{aitoa.plot.progress.on.instance}
\title{Plot the Progress of a Set of Algorithms on One Instance}
\usage{
aitoa.plot.progress.on.instance(
  results.dir = ".",
  algorithms,
  instance,
  instance.name = instance,
  time.column = c("t", "fes"),
  max.time = NA_integer_,
  algorithm.colors = aitoa.distinct.colors(length(algorithms)),
  instance.limit = NA_integer_,
  instance.limit.name = NA_character_,
  ...
)
}
\arguments{
\item{results.dir}{the directory where the results can be loaded from}

\item{algorithms}{a list of algorithsm, the \code{names} of which (if
provided) are used for the legend}

\item{time.column}{the time column}

\item{max.time}{an optional maximal time limit}

\item{algorithm.colors}{a character vector of the same length as
\code{algorithms} providing the colors to be used for the algorithms}

\item{instance.limit}{an opional quality limit to be plotted as horizontal
line}

\item{instance.limit.name}{the optional name of the quality limit}

\item{...}{parameters to be passed to \link[graphics]{par}}
}
\description{
Plot the progress of a set of algorithms on one instance.
}
