% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_progress_on_multiple_instances.R
\name{aitoa.plot.progress.on.multiple.instances}
\alias{aitoa.plot.progress.on.multiple.instances}
\title{Plot the Progress for a Set of Algorithms on a Set of Problem
  Instances}
\usage{
aitoa.plot.progress.on.multiple.instances(
  results.dir,
  algorithms,
  instances,
  time.column = c("t", "fes"),
  max.time = NA_integer_,
  instances.limits = NA_integer_,
  algorithm.colors = aitoa.distinct.colors(length(algorithms)),
  ...
)
}
\arguments{
\item{results.dir}{the directory where the results can be loaded from}

\item{algorithms}{the list of algorithm IDs. The \code{names} of this list,
if set, will be used in the legends.}

\item{instances}{the list of instance IDs. The \code{names} of this list, if
set, will be used in the legend}

\item{time.column}{the time dimension, either \code{t} or \code{fes}}

\item{max.time}{an optional limit for the time}

\item{instances.limits}{an optional vector of lower bounds or best-known
solutions for the instances (the \code{names} of the vector will be used in
the legend)}

\item{algorithm.colors}{the colors to be used for the different algorithms}

\item{...}{parameters to be passed to \link[graphics]{par}}
}
\description{
Plot how a set of algorithms progress over a set of problem
  instances. For each instance, one diagram is plotted. The diagrams are
  arranged one by one from in a vertical row.
}
