% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stat_over_param.R
\name{aitoa.plot.stat.over.param}
\alias{aitoa.plot.stat.over.param}
\title{Plot a Statistic over the Values of a Given Parameter}
\usage{
aitoa.plot.stat.over.param(
  end.result.stats,
  algorithm.template,
  algorithm.args,
  instances,
  statistic,
  instance.colors = aitoa.distinct.colors(length(instances)),
  instance.lty = .default.lty,
  instance.lwd = .default.lwd,
  instance.pch = NULL,
  legend.pos = "topright",
  legend.cex = .legend.cex,
  legend.bg = .legend.bg,
  mgp = .default.mgp,
  tck = .default.tck,
  cex = .default.cex,
  mar = .default.mar.without.labels,
  divide.by = NULL,
  ...
)
}
\arguments{
\item{end.result.stats}{the end result statistics frame}

\item{algorithm.template}{the algorithm template: a string where \code{$arg}
will be replaced with the parameter value and which will then be matched
with the \code{algorithm} column in the data frame}

\item{algorithm.args}{the values of the argument}

\item{instances}{a potentially named list of instances; if names are given,
the names are diplayed in the legend, otherwise the instance strings}

\item{statistic}{the statistic to plot}

\item{instance.colors}{the colors to use for the instances}

\item{instance.lty}{the line types to be used for the instances}

\item{instance.lwd}{the line widths to be used for the instances}

\item{instance.pch}{the symbol to be used for the instances}

\item{legend.pos}{the legend position}

\item{legend.cex}{the legend character size}

\item{legend.bg}{the legend background}

\item{mgp}{the mgp value}

\item{tck}{the tck value}

\item{cex}{the base character scaling}

\item{mar}{the margin value}

\item{divide.by}{optionally, a per-instance value to divide the statistic by}

\item{...}{parameters to be passed to \link[graphics]{plot}}
}
\description{
Plot an end statistic over the values of a given parameter.
}
